refactor(authn+db+api): rewrite funcnctions to use tokenIdentifier, optimized player creation

+++ convex/mutations.ts
@@ -1,15 +1,16 @@
 import { v } from "convex/values";
 import { internalMutation, mutation } from "./_generated/server";
+import { internal } from "./_generated/api";
 import { api } from "./_generated/api";
 
 export const createPlayer: ReturnType<typeof internalMutation> = internalMutation({
   args: {
-    clerk_id: v.string(),
+    tokenIdentifier: v.string(),
     nickname: v.string(),
   },
-  handler: async (ctx, { clerk_id, nickname }) => {
+  handler: async (ctx, { tokenIdentifier, nickname }) => {
     const player = await ctx.db.insert("player", {
-      clerk_id: clerk_id,
+      tokenIdentifier: tokenIdentifier.split('|').reverse().join('|'),
       nickname: nickname,
       imgURL: "",
       xp: 0,
@@ -28,14 +29,31 @@ export const createPlayer: ReturnType<typeof internalMutation> = internalMutatio
   },
 });
 
+export const createNewPlayer: ReturnType<typeof mutation> = mutation({
+  args: {},
+  handler: async (ctx) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("User must be authenticated to create a player");
+    }
+    const existing = await ctx.runQuery(api.queries.getPlayerByTokenIdentifier, {});
+    if (existing) {
+      throw new Error("Player already exists for this user");
+    }
+    const player = await ctx.runMutation(internal.mutations.createPlayer, { tokenIdentifier: identity?.tokenIdentifier as string, nickname: identity?.nickname as string });
+    return player;
+  },
+});
 
 
 export const deletePlayer: ReturnType<typeof mutation> = mutation({
-  args: {
-    clerk_id: v.string(),
-  },
-  handler: async (ctx, args) => {
-    const player = await ctx.runQuery(api.queries.getPlayerByClerkId, { clerk_id: args.clerk_id });
+  args: {},
+  handler: async (ctx) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("User must be authenticated to delete a player");
+    }
+    const player = await ctx.runQuery(api.queries.getPlayerByTokenIdentifier, { tokenIdentifier: identity?.tokenIdentifier as string });
     const deleted = await ctx.db.delete("player", player._id);
     return deleted;
   },
